Refactor using Flask Blueprints
Instead of one massive routes.py file, I am now using
Flask Blueprints to divide the routes into their
respective locations. This required editing several
pages all at once in order to get proper functioning,
which is why most files are being updated with this push.
