--- .flaskenv
@@ -1,2 +1,0 @@
-FLASK_APP=run.py
-FLASK_ENV=development
