+++ .gitignore
@@ -16,6 +16,7 @@ __pycache__/
 *$py.class
 
 # Flask #
+.flaskenv
 instance/
 .webassets-cache
