+++ app/templates/add_dessert.html
@@ -2,7 +2,7 @@
 {% block title %}Delicious Dessert in Progress | {% endblock %}
 {% block content %}
 
-<form action="{{ url_for('add_dessert_toDB') }}" method="POST">
+<form action="{{ url_for('recipes.add_dessert_toDB') }}" method="POST">
 
     <div class="container-75">
