+++ app/templates/index.html
@@ -53,26 +53,26 @@
             <!-- Quick Links -->
             <div class="center-align quick-links">
                 <h5 class="bold purple-text font-special upper">Quick Links</h5>
-                <a href="{{ url_for('view_desserts') }}" class="btn bold purple white-text text-shadow-2">
+                <a href="{{ url_for('recipes.view_desserts') }}" class="btn bold purple white-text text-shadow-2">
                     <i class="fas fa-search material-icons left" aria-hidden="true"></i> Search
                 </a>
                 <!-- only if user in session -->
                 {% if session.user %}
-                <a href="{{ url_for('profile', username=session['user']) }}" class="btn bold blue white-text text-shadow-2">
+                <a href="{{ url_for('users.profile', username=session['user']) }}" class="btn bold blue white-text text-shadow-2">
                     <i class="fas fa-user-circle material-icons left" aria-hidden="true"></i> Profile
                 </a>
-                <a href="{{ url_for('add_dessert') }}" class="btn bold green white-text text-shadow-2">
+                <a href="{{ url_for('recipes.add_dessert') }}" class="btn bold green white-text text-shadow-2">
                     <i class="fas fa-plus material-icons left" aria-hidden="true"></i> Add
                 </a>
                 <!-- if no user in session -->
                 {% else %}
-                <a href="{{ url_for('login') }}" class="btn bold pink lighten-2 white-text text-shadow-2">
+                <a href="{{ url_for('users.login') }}" class="btn bold pink lighten-2 white-text text-shadow-2">
                     <i class="fas fa-user-plus material-icons left" aria-hidden="true"></i> Log In | Register
                 </a>
                 {% endif %}
                 <!-- get random recipe -->
                 {% for recipe in random_recipe %}
-                <a href="{{ url_for('view_dessert', recipe_id=recipe._id, slugUrl=recipe.recipe_slug) }}" class="btn bold red white-text text-shadow-2">
+                <a href="{{ url_for('recipes.view_dessert', recipe_id=recipe._id, slugUrl=recipe.recipe_slug) }}" class="btn bold red white-text text-shadow-2">
                     <i class="fas fa-random material-icons left" aria-hidden="true"></i> Random
                 </a>
                 {% endfor %}
@@ -86,7 +86,7 @@
                 {% for recipe in carousel %}
                 <div class="carousel-item">
                     <div class="carousel-fixed-item center text-shadow-1">
-                        <a href="{{ url_for('view_dessert', recipe_id=recipe._id, slugUrl=recipe.recipe_slug) }}">
+                        <a href="{{ url_for('recipes.view_dessert', recipe_id=recipe._id, slugUrl=recipe.recipe_slug) }}">
                             <h4 class="bold font-emilys-candy mark-purple text-shadow-2">
                                 {{ recipe.recipe_name|truncate(50, True)|safe }}
                             </h4>
