+++ app/templates/update_dessert.html
@@ -2,7 +2,7 @@
 {% block title %}Tweaking the Taste of {{ recipe.recipe_name }} | {% endblock %}
 {% block content %}
 
-<form action="{{ url_for('update_dessert_toDB', recipe_id=recipe._id) }}" method="POST">
+<form action="{{ url_for('recipes.update_dessert_toDB', recipe_id=recipe._id) }}" method="POST">
 
     <div class="container-75">
