+++ app/templates/view_desserts.html
@@ -120,7 +120,7 @@
                             <button id="search_btn" class="btn waves-effect blue waves-light text-shadow-2 bold upper" type="submit" disabled>Search
                                 <i class="fas fa-search material-icons left" aria-hidden="true"></i>
                             </button>
-                            <a class="btn waves-effect waves-purple pink lighten-2 text-shadow-2 bold upper tooltipped" data-position="top" data-tooltip="Reset All" href="{{ url_for('view_desserts') }}">
+                            <a class="btn waves-effect waves-purple pink lighten-2 text-shadow-2 bold upper tooltipped" data-position="top" data-tooltip="Reset All" href="{{ url_for('recipes.view_desserts') }}">
                                 <i class="fas fa-undo-alt material-icons" aria-hidden="true"></i>
                             </a>
                         </div>
